[renader] draw_state update, macros modularization
